Refactor window show/hide hooks and add bypass logic
Refactored the window show/hide hook mechanism to allow hooks to intercept and optionally bypass the original show/hide implementation. Added HasWillShowHook and HasWillHideHook checks, and introduced HandleWillShow/HandleWillHide methods. Updated Windows-specific code to support skipping the original implementation if a hook handles the event, and provided methods to call the original show/hide logic if needed.

+++ src/capi/window_manager_c.cpp
@@ -153,7 +153,7 @@ native_window_list_t native_window_manager_get_all(void) {
       return result;
     }
 
-    result.count = windows.size();
+    result.count = static_cast<long>(windows.size());
 
     for (size_t i = 0; i < windows.size(); ++i) {
       result.windows[i] = CreateNativeWindowHandle(windows[i]);
